Day 22 - Finish ft_printf

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mboujama <mboujama@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/23 10:45:05 by mboujama          #+#    #+#             */
-/*   Updated: 2023/12/24 13:19:27 by mboujama         ###   ########.fr       */
+/*   Updated: 2023/12/25 14:03:11 by mboujama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int		ft_printf(const char *str, ...);
 void	ft_putchar(char s, int *len);
 void	ft_putnbr(long long n, int *len);
 void	ft_putstr(char *s, int *len);
-void	ft_puthexa(unsigned long nb, char c, int *len);
-void	ft_putpointer(unsigned long nb, int *len);
+void	ft_puthexa(unsigned int nb, char c, int *len);
+void	ft_putpointer(unsigned int nb, int *len);
 
 #endif
